refactor: change discount from int to BigDecimal

+++ src/main/java/com/hmigl/cdc/coupon/Coupon.java
@@ -8,6 +8,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -15,7 +16,7 @@ public class Coupon {
     private @Id @GeneratedValue Long id;
 
     private @NotBlank String code;
-    private @NotNull @Positive Integer discountPercentage;
+    private @NotNull @Positive BigDecimal discountPercentage;
     private @NotNull @Future LocalDate expirationDate;
 
     @Deprecated
@@ -23,7 +24,7 @@ public class Coupon {
 
     public Coupon(
             @NotBlank String code,
-            @NotNull @Positive Integer discountPercentage,
+            @NotNull @Positive BigDecimal discountPercentage,
             @NotNull @Future LocalDate expirationDate) {
         this.code = code;
         this.discountPercentage = discountPercentage;
